Bump version, move changes to index.js, and add notes on breaking changes to README.md
