+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "knackhq-client",
-  "version": "0.1.1",
-  "description": "KnackHQ API Client Module",
+  "version": "1.0.0",
+  "description": "Knack API Client Module",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
